Add files via upload
This is an  Mobile app development project about memory card game. This is an simple project The game is made to help users improve their memory skills. Players need to match pairs of cards by remembering their positions. It is easy to play and fun for all ages. The project is developed in andriod studio using kotlin language.